add support for user-input movements
The built-in movement system was used for additional flavor text and soft suggestions for workouts, but ended up more of a hindrance when the flavor text offered little value, and extra work had to be put in to creating an interface that reflects certain movements.

In favor of more user freedom and simpler design, there is no longer a database of supported movements, but instead a workout is comprised of user-inputted movements - meaning that any user can write down any exercise as part of their workout.